+++ src/sequential/nilsson-forward8.cpp
@@ -35,16 +35,16 @@ typedef unsigned char* string;
 #define FALSE 0
 #define MAX(a, b) ((a) > (b) ? (a) : (b))
 
-typedef int boolean;
-typedef int character;
+typedef long boolean;
+typedef long character;
 
 /******************* A simple memory allocator ********************/
 
 typedef struct {
 	void *block[MAXBLOCKS];
-	int allocnr;
-	int nr;
-	int blocksize;
+        long allocnr;
+        long nr;
+        long blocksize;
 	void *current, *first, *last;
 } memory;
 
@@ -53,7 +53,7 @@ typedef struct {
    blocksize * MAXBLOCKS memory slots each consisting of
    elemsize bytes.
 */
-static inline void initmem(memory *m, int elemsize, int blocksize)
+static inline void initmem(memory *m, long elemsize, long blocksize)
 {
    m->blocksize = MAX(blocksize, 1000) * elemsize;
    m->block[0] = malloc(m->blocksize);
@@ -62,7 +62,7 @@ static inline void initmem(memory *m, int elemsize, int blocksize)
    m->current = m->first = m->last = NULL;
 }
 
-static inline void *allocmem(memory *m, int elemsize)
+static inline void *allocmem(memory *m, long elemsize)
 {
    if (m->current < m->last)
       m->current = (void *) ((char *) (m->current) + elemsize);
@@ -79,7 +79,7 @@ static inline void *allocmem(memory *m, int elemsize)
    return m->current;
 }
 
-static inline void *deallocmem(memory *m, int elemsize)
+static inline void *deallocmem(memory *m, long elemsize)
 {
    if (m->current > m->first)
       m->current = (void *) ((char *) (m->current) - elemsize);
@@ -103,7 +103,7 @@ static inline void resetmem(memory *m)
 
 static inline void freemem(memory *m)
 {
-   int i;
+   long i;
 
    for (i = 0; i <= m->allocnr; i++)
       free(m->block[i]);
@@ -117,10 +117,10 @@ typedef struct listrec *list;
 struct listrec {
    string str;
    list next;
-   int length;
+   long length;
 };
 
-static inline int scmp(unsigned char *s1, unsigned char *s2)
+static inline long scmp(unsigned char *s1, unsigned char *s2)
 {
     while( *s1 != '\0' && *s1 == *s2 )
         s1++, s2++;
@@ -129,7 +129,7 @@ static inline int scmp(unsigned char *s1, unsigned char *s2)
 
 /* Insertion sort for linked lists of character strings.
    The strings all have a common prefix of length p. */
-static inline list Insertsort(list r, list *tail, int p)
+static inline list Insertsort(list r, list *tail, long p)
 {
    list fi, la, t;
 
@@ -175,7 +175,7 @@ struct grouprec {
 
 struct bucketrec {
    list head, tail; /* a list of elements */
-   int size;        /* list length */
+   long size;        /* list length */
    group tag;       /* group tag */
    bucket next;     /* next bucket item */
 };
@@ -188,7 +188,7 @@ static memory bucketmem[1];
    be inserted the list is just appended, otherwise a new bucket
    is created. */
 static void intobucket(bucket *b, list head, list tail,
-                       int size, group g)
+                       long size, group g)
 {
    bucket btemp = *b, newb;
 
@@ -215,13 +215,13 @@ static void intobucket(bucket *b, list head, list tail,
    blocks consisting of strings that have a common character in
    position pos. Furthermore, a group that is not split during this
    phase is left behind and not put into a bucket. */
-static void intobuckets(group g, bucket b[], int pos)
+static void intobuckets(group g, bucket b[], long pos)
 {
    group prevg;
    character ch, prevch;
    boolean split;
    list tail, tailn;
-   int size;
+   long size;
 
    resetmem(bucketmem);
    for (prevg = g, g = g->nextunf ; g; g = g->nextunf) {
@@ -277,7 +277,7 @@ static void intogroup(group g, list head, list tail, boolean finis)
 /* Traverse the buckets and put the elements back into their groups.
    Split the groups and mark all finished groups.
    The elements are moved in blocks. */
-static void intogroups(bucket b[], int pos)
+static void intogroups(bucket b[], long pos)
 {
    character ch;
    bucket s;
@@ -313,11 +313,11 @@ static list collect(group g)
    return head;
 }
 
-static inline list forward1(list t, int n)
+static inline list forward1(list t, long n)
 {
    static bucket b[CHARS];   /* buckets */
    group g, g2;              /* groups */
-   int pos = 0;              /* pos in string */
+   long pos = 0;              /* pos in string */
 
    if (n<2) return t;
